crqMI {crqMI}	R Documentation

Functions to fit censored quantile regression models with covariates subject to detection limits

Description

The method is developed using some techniques in Peng and Huang (2008) and Wang and Feng (2012).

Usage


crqMI(Y, Z, X, d, delta.v, delta.t, Imp = NULL, M = 10,
tau=c(0.25,0.5,0.75), ctype="left", seed=TRUE,
pi.fam="logistic",se=FALSE, Boot=100, shrinkage=FALSE)

Arguments

Y
survival outcomes.

Z
observations of censored covariates.

X
fully observed covariates.

d
detection limits for censored covariates Z.

delta.v
censoring indicators for Z. 1 indicates observed otherwise censored.

delta.t
censoring indicators for Y. 1 indicates observed otherwise censored.

Imp
imputation matrix. Default is Imp=NULL, regarding (X,log(Y),delta.c) as the impuation model.

M
the number of imputation times

tau
the quantile(s) at which the model is to be estimated.

ctype
censoring type(s) of censored covariate(s), left or right censored. It can be a character or a vector of characters. By default ctype="left", regarding all censored covariates as left-censored.

seed
If seed=TRUE, see seed in the computation.

pi.fam
Default is pi.fam="logistic", using logistic regression to estimate the conditional censoring probability Pr(delta.v=0|imp). Another choice is pi.fam="probit"

se
If se=TRUE, returning the estimates of standard error (SE) based on the Bootstrap method. Default is se=FALSE

Boot
the number of bootstrap resamples to estimate SE. Default is Boot=100

shrinkage
If shrinkage=TRUE, returning the shrinkage estimate. Default is shrinkage=FALSE

Value

estimate
a list of estimates obtained from complete case analysis (CC), naive method ignoring censoring of covariates (filld), the proposed method (MI) and its shrinkage version (SMI)

se
a list of estimated stardard errors obtained from complete case analysis (CC), naive method ignoring censoring of covariates (filld), the proposed method (MI) and its shrinkage version (SMI)

Author(s)

Yu Tonghui, Xiang Liming and Wang Huixia Judy.

References

Peng, L. and Huang, Y. (2008). Survival analysis with quantile regression models, J. Am. Stat. Assoc., 103, 637-649.

Wang, H.J. and Feng, X. (2012) Multiple imputation for m-regression with censored covariates, J. Am. Stat. Assoc., 104, 1117-1128.

See Also

crq

Examples


library(crqMI)
set.seed(1234)
datasim=sim(n=100)
data = datasim$obs.data

tau=c(0.3,0.5)

############  if SEs are not required #####
out=crqMI(Y= data$Y, Z= data$Z, X=data$X, d= datasim$d,
          delta.v=data$delta.v, delta.t=data$delta.t,  Imp=NULL,
          M = 10, tau=tau, ctype= datasim$ctype,seed=T,
          pi.fam="logistic",se=FALSE,shrinkage=TRUE)

MI=out$estimate$MI  ## MI estimates
CC=out$estimate$CC  ## CC estimates
true.par =datasim$true.par[,paste(tau)]
print(round(cbind(MI-true.par,CC-true.par),3)) # the estimation errors

############  need to compute SEs and shrinkage estimates#####
out=crqMI(Y= data$Y, Z= data$Z, X=data$X, d= datasim$d,
          delta.v=data$delta.v, delta.t=data$delta.t,  Imp=NULL,
          M = 10, tau=tau, ctype= datasim$ctype,seed=T,
          pi.fam="logistic",se=TRUE,Boot=100,shrinkage=TRUE)


MI=out$estimate$MI  ## MI estimates
CC=out$estimate$CC  ## CC estimates
SMI=out$estimate$SMI ## shrinkage estimates
## estimation error and SE
true.par =datasim$true.par[,paste(tau)]
print(round(cbind(MI-true.par,SMI-true.par,CC-true.par),3))
print(round(cbind(out$se$MI,out$se$SMI,out$se$CC),3))



[Package crqMI version 0.1.0 Index]